chore(tests): move check_requests from `encryption_sync_service` to `sliding_sync` helper file
